some updates on calcs

+++ app/src/main/java/com/example/frames_demo/datadog/frame_calculation/FrameRecord.kt
@@ -4,7 +4,7 @@ import java.util.concurrent.TimeUnit
 
 internal data class SlowFrameRecord(
     val startTimestampNs: Long,
-    var durationNs: Long,
+    var durationNs: Float,
 ) {
     override fun toString(): String {
         return "${durationNs / NS_IN_MS}ms"
